Split freeing of parent object from children
The 'DhcpStructure a' dictionary object has always had a function named
'freeDhcp'. For most structures, this would recursively free all child
structures before freeing the "parent" pointer supplied to the
function.

There is at least one case where this final step was not desired. The
'hostInfo :: DhcpStructure HOST_INFO' declaration didn't free the parent
because on the C side its structure was inlined within a
'ClientInfoV4'. There was another case with 'DhcpStructure (LengthBuffer
a)' where I should have been freeing the parent pointer, but was not.

The 'DhcpStructure a' dictionary now contains a member named
'freeDhcpChildren' which makes it clear that the main pointer will not
be freed. There is a separate function outside the dictionary named
'freeDhcp' which simply calls 'freeDhcpChildren' before freeing the main
pointer.

Most call sites were left as before, and code dealing with 'HOST_INFO'
isn't a special case.