LengthArray variants clean up buffer spine
LengthArrays are created by specifying a "flavor" of array. Available
options include baseArray, basicArray, and ptrArray. The difference has
to do with how the underlying buffer is layed out in memory. Previously
the baseArray and basicArray variants were failing to free the spine of
the buffer before the lengthArray itself was freed.