enum functions ignore read and total out vars
The enumSubnetElementsV4 and enumSubnetClientsV4 are simplified to
ignore the elementsRead and elementsTotal out vars. MSDN documentation
advices that these be referred to deciding whether to call the function
additional times with a resumeHandle.

In my testing I've found that the elementsTotal function always returns
0x7fffffff until the final call. elementsRead does return a useful
value, but not for my use case. In either case, the only thing to watch
for when deciding whether to call the enum function additional times is
the return value.