replace 'freeptr' with "scrubbing" family
I replaced the 'freeptr' function with a more general "scrubbing" family
of functions. These perform an action over the pointer within a doubly
referenced pointer if and only if no access violation would occur. It
then returns the result of that action wrapped in a 'Maybe'. A nothing
indicates that the embedded pointer had a value of 0.